feat(server): ✨ | restructure options handling in server
introduces a new options parsing structure by reorganizing source files into an options_parser directory.
adds new parser functionality with dedicated processing logic.
renames and relocates existing option handling files to improve code organization and clarity.

THIS COMMIT SHOULD BE SPLIT.
