+++ Makefile
@@ -62,15 +62,16 @@ INCLUDES_SERVER = ${INCLUDES}server/
 
 # Sources
 SRC_SERVER = $(SRCDIR_SERVER)main.c	\
-			 $(SRCDIR_SERVER)option_parser.c	\
-			 $(SRCDIR_SERVER)options/port.c	\
-			 $(SRCDIR_SERVER)options/width.c	\
-			 $(SRCDIR_SERVER)options/height.c	\
-			 $(SRCDIR_SERVER)options/teams.c	\
-			 $(SRCDIR_SERVER)options/clients.c	\
-			 $(SRCDIR_SERVER)options/frequency.c	\
-			 $(SRCDIR_SERVER)options/help.c	\
-			 $(SRCDIR_SERVER)options/debug.c	\
+			 $(SRCDIR_SERVER)options_parser/parser.c	\
+			 $(SRCDIR_SERVER)options_parser/processor.c	\
+			 $(SRCDIR_SERVER)options_parser/options/port.c	\
+			 $(SRCDIR_SERVER)options_parser/options/width.c	\
+			 $(SRCDIR_SERVER)options_parser/options/height.c	\
+			 $(SRCDIR_SERVER)options_parser/options/teams.c	\
+			 $(SRCDIR_SERVER)options_parser/options/clients.c	\
+			 $(SRCDIR_SERVER)options_parser/options/frequency.c	\
+			 $(SRCDIR_SERVER)options_parser/options/help.c	\
+			 $(SRCDIR_SERVER)options_parser/options/debug.c	\
 
 # Objects
 OBJ_SERVER = $(SRC_SERVER:.c=.o)
