+++ src/server/options_parser/options/clients.c
@@ -5,7 +5,7 @@
 ** Clients Option
 */
 
-#include "options.h"
+#include "options_parser/parser.h"
 #include <stdbool.h>
 #include <stdlib.h>
 #include <stdio.h>
