+++ src/server/options_parser/options/frequency.c
@@ -5,7 +5,7 @@
 ** Frequency Option
 */
 
-#include "options.h"
+#include "options_parser/parser.h"
 #include <stdio.h>
 #include <stdlib.h>
