+++ src/server/options_parser/options/height.c
@@ -5,7 +5,7 @@
 ** Height Option
 */
 
-#include "options.h"
+#include "options_parser/parser.h"
 #include <stdio.h>
 #include <stdlib.h>
