+++ src/server/options_parser/options/teams.c
@@ -5,7 +5,7 @@
 ** Teams Option
 */
 
-#include "options.h"
+#include "options_parser/parser.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
