+++ src/server/options_parser/options/width.c
@@ -5,7 +5,7 @@
 ** Width Option
 */
 
-#include "options.h"
+#include "options_parser/parser.h"
 #include <stdio.h>
 #include <stdlib.h>
