+++ src/server/options_parser/parser.c
@@ -5,7 +5,7 @@
 ** Option Parser
 */
 
-#include "options.h"
+#include "options_parser/parser.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
